small modifications to resolve recent conversations
