+++ mmdet/models/layers/__init__.py
@@ -15,14 +15,16 @@ from .positional_encoding import (LearnedPositionalEncoding,
                                   SinePositionalEncoding)
 from .res_layer import ResLayer, SimplifiedBasicBlock
 from .se_layer import ChannelAttention, DyReLU, SELayer
-from .transformer import (MLP, DeformableDetrTransformerDecoder,
+from .transformer import (MLP, CdnQueryGenerator,
+                          DeformableDetrTransformerDecoder,
                           DeformableDetrTransformerDecoderLayer,
                           DeformableDetrTransformerEncoder,
                           DeformableDetrTransformerEncoderLayer,
                           DetrTransformerDecoder, DetrTransformerDecoderLayer,
                           DetrTransformerEncoder, DetrTransformerEncoderLayer,
-                          DynamicConv, PatchEmbed, PatchMerging,
-                          inverse_sigmoid, nchw_to_nlc, nlc_to_nchw)
+                          DinoTransformerDecoder, DynamicConv, PatchEmbed,
+                          PatchMerging, inverse_sigmoid, nchw_to_nlc,
+                          nlc_to_nchw)
 
 __all__ = [
     'fast_nms', 'multiclass_nms', 'mask_matrix_nms', 'DropBlock',
@@ -37,5 +39,6 @@ __all__ = [
     'DetrTransformerEncoder', 'DetrTransformerDecoder',
     'DeformableDetrTransformerEncoder', 'DeformableDetrTransformerDecoder',
     'DeformableDetrTransformerEncoderLayer',
-    'DeformableDetrTransformerDecoderLayer', 'DinoTransformerDecoder'
+    'DeformableDetrTransformerDecoderLayer', 'DinoTransformerDecoder',
+    'CdnQueryGenerator'
 ]
